Add support for Yoshino platform

+++ runme_1unpack64.sh
@@ -3,6 +3,7 @@ if [ -d work ]; then
     rm -rf ./work
 fi
 
+# ELF format (SIN v1 ~ SIN v3)
 if [ ! -e kernel.elf ]; then
     if [ -e kernel*.elf ]; then
         mv kernel*.elf kernel.elf
@@ -13,9 +14,34 @@ if [ ! -e kernel.elf ]; then
     fi
 fi
 
+# Android image format (SIN v4 ~ )
+if [ ! -e kernel.bin ]; then
+    if [ -e kernel*.bin ]; then
+        mv kernel*.bin kernel.bin
+    fi
+fi
+
 mkdir work
-mv kernel.elf work/kernel.sin
-cp tools/unpack-kernelsin64.pl work/unpack-kernelsin64.pl
+
+if [ -e kernel.elf ]; then
+	mv kernel.elf work/kernel.sin
+	cp -a tools/unpack-kernelsin64.pl work/unpack-kernelsin64.pl
+elif [ -e kernel.bin ]; then
+	mv kernel.bin work/kernel.bin
+fi
+
+
 cd work
-chmod 777 unpack-kernelsin64.pl
-./unpack-kernelsin64.pl kernel.sin
+
+if [ -e kernel.elf ]; then
+	chmod 777 unpack-kernelsin64.pl
+	./unpack-kernelsin64.pl kernel.sin
+elif [ -e kernel.bin ]; then
+	../tools/unpackbootimg -i kernel.bin
+	mkdir kernel.sin-ramdisk
+	mv kernel.bin-ramdisk.gz kernel.sin-ramdisk/kernel.bin-ramdisk.gz
+	cd kernel.sin-ramdisk
+	gunzip -c kernel.bin-ramdisk.gz | cpio -i
+	rm -f kernel.bin-ramdisk.gz
+fi
+
